+++ app/Http/Controllers/admin/UserController.php
@@ -12,7 +12,7 @@ use Illuminate\Support\Facades\Cache;
 use App\Http\Requests\User\UpdateUser;
 use App\Repositories\users\userrepository;
 
-class UserController extends Controller 
+class UserController extends Controller
 {
     public $userrepository;
 
@@ -24,16 +24,23 @@ class UserController extends Controller
 
     public function index()
     {
-            $userlist = $this->userrepository->index();
-            return view('admin/users/index', compact('userlist'));
+
+        if (Gate::denies('show_user')) {
+            abort(403);
+        }
+        $userlist = $this->userrepository->index();
+        return view('admin/users/index', compact('userlist'));
     }
 
 
     public function store(StoreUser $request)
     {
+        if (Gate::denies('create_user')) {
+            abort(403);
+        }
         $data = request()->except("_token");
         $result = $this->userrepository->store($data);
-       
+
         return response()->json([
             'status' => '200',
             'message' => 'Thêm mới thành công',
@@ -41,32 +48,40 @@ class UserController extends Controller
     }
     public function edit($id)
     {
+        if (Gate::denies('show_user')) {
+            abort(403);
+        }
         $getUser = $this->userrepository->edit($id);
-        $getRole = DB::table('role_users')->select('role_id')->where('user_id',$id)->get();
+        $getRole = DB::table('role_users')->select('role_id')->where('user_id', $id)->get();
         return response()->json(
-            compact('getUser','getRole')
+            compact('getUser', 'getRole')
         );
     }
 
 
-    public function update( UpdateUser $request)
-    {       
-            $user = $this->userrepository->edit($request['id']);
-            $data = request()->except("_token",'id');
-            $this->userrepository->update($user, $data);
-            return response()->json([
-                'status' => '200',
-                'message' => 'Cập nhật thành công',
-            ]);
-        
+    public function update(UpdateUser $request)
+    {
+        if (Gate::denies('edit_user')) {
+            abort(403);
+        }
+        $user = $this->userrepository->edit($request['id']);
+        $data = request()->except("_token", 'id');
+        $this->userrepository->update($user, $data);
+        return response()->json([
+            'status' => '200',
+            'message' => 'Cập nhật thành công',
+        ]);
     }
 
     public function delete(Request $request)
     {
+        if (Gate::denies('delete_user')) {
+            abort(403);
+        }
         $this->userrepository->delete($request);
-       return response()->json([
-        'status' => '100',
-        'message' => 'Xóa thành công',
-       ]);
+        return response()->json([
+            'status' => '100',
+            'message' => 'Xóa thành công',
+        ]);
     }
 }
